deps: bump emulator version to latest and optimize startup
Updates the dependencies for testing using the emulator to the latest. This also includes
always using the latest version of the emulator for tests.
The startup time of the emulator based tests is optmized by adding a wait-for for the
listening port of the emulator, instead of relying on a 10 second startup-wait-time and
an error being returned for a non-existing URL.