deps: update Liquibase test harness to 1.0.9
Bumps the Liquibase test harness version to 1.0.9 and aligns all other relevant test
dependencies with the versions that are required for that. Note that the test harness
version 1.0.9 no longer works with Java 8, and requires at least Java 11 to run. This
change therefore also adds a build profile that is only activated for Java 8 and that
excludes running the Liquibase test harness tests. The tests are run on Java 11 and
higher. This is also done in a separate profile to prevent dependencies from being
added to the build when the tests are executed using Java 8.

The provider itself is still usable with Java 8.