fix: support snapshotting indexes with include columns (#163)
* fix: support snapshotting indexes with include columns

Adds support for taking a snapshot of an index with a STORING column clause.
These failed because Liquibase assumes that the ordinal position of a column
would always be non-null. In Cloud Spanner, index columns in a STORING clause
have a null value for their ordinal position.

Note that the STORING clause is not generated by Liquibase, as Liquibase itself
does nothing with that part of an index. This could be added to a custom index
generator for Cloud Spanner in a separate PR.

Fixes #159

* chore: remove unused references