feat: support bit-reversed sequences (#214)
* feat: support bit-reversed sequences

Adds support for creating and dropping bit-reversed sequences. Renaming
a sequence is still not supported.

The minValue and maxValue properties in Liquibase do not have a
corresponding property in Cloud Spanner, but at the same time, Cloud
Spanner has skip_range_min and skip_range_max, which are not available
in Liquibase. The minValue van maxValue properties in Liquibase are
therefore mapped to the skip-range values in Cloud Spanner.

* test: add integration test